# zstd: Skip random data faster