--- a/zstd/encoder_test.go
+++ b/zstd/encoder_test.go
@@ -723,3 +723,103 @@ func BenchmarkEncoder_EncodeAllPi(b *testing.B) {
 		}
 	}
 }
+
+func BenchmarkRandomEncodeAllFastest(b *testing.B) {
+	rng := rand.New(rand.NewSource(1))
+	data := make([]byte, 10<<20)
+	for i := range data {
+		data[i] = uint8(rng.Intn(256))
+	}
+	enc, _ := NewWriter(nil, WithEncoderLevel(SpeedFastest), WithEncoderConcurrency(1))
+	dst := enc.EncodeAll(data, nil)
+	wantSize := len(dst)
+	b.ResetTimer()
+	b.ReportAllocs()
+	b.SetBytes(int64(len(data)))
+	for i := 0; i < b.N; i++ {
+		dst := enc.EncodeAll(data, dst[:0])
+		if len(dst) != wantSize {
+			b.Fatal(len(dst), "!=", wantSize)
+		}
+	}
+}
+
+func BenchmarkRandomEncodeAllDefault(b *testing.B) {
+	rng := rand.New(rand.NewSource(1))
+	data := make([]byte, 10<<20)
+	for i := range data {
+		data[i] = uint8(rng.Intn(256))
+	}
+	enc, _ := NewWriter(nil, WithEncoderLevel(SpeedDefault), WithEncoderConcurrency(1))
+	dst := enc.EncodeAll(data, nil)
+	wantSize := len(dst)
+	b.ResetTimer()
+	b.ReportAllocs()
+	b.SetBytes(int64(len(data)))
+	for i := 0; i < b.N; i++ {
+		dst := enc.EncodeAll(data, dst[:0])
+		if len(dst) != wantSize {
+			b.Fatal(len(dst), "!=", wantSize)
+		}
+	}
+}
+
+func BenchmarkRandomEncoderFastest(b *testing.B) {
+	rng := rand.New(rand.NewSource(1))
+	data := make([]byte, 10<<20)
+	for i := range data {
+		data[i] = uint8(rng.Intn(256))
+	}
+	wantSize := int64(len(data))
+	enc, _ := NewWriter(ioutil.Discard, WithEncoderLevel(SpeedFastest))
+	n, err := io.Copy(enc, bytes.NewBuffer(data))
+	if err != nil {
+		b.Fatal(err)
+	}
+	if n != wantSize {
+		b.Fatal(n, "!=", wantSize)
+	}
+	b.ResetTimer()
+	b.ReportAllocs()
+	b.SetBytes(wantSize)
+	for i := 0; i < b.N; i++ {
+		enc.Reset(ioutil.Discard)
+		n, err := io.Copy(enc, bytes.NewBuffer(data))
+		if err != nil {
+			b.Fatal(err)
+		}
+		if n != wantSize {
+			b.Fatal(n, "!=", wantSize)
+		}
+	}
+}
+
+func BenchmarkRandomEncoderDefault(b *testing.B) {
+	rng := rand.New(rand.NewSource(1))
+	data := make([]byte, 10<<20)
+	for i := range data {
+		data[i] = uint8(rng.Intn(256))
+	}
+	wantSize := int64(len(data))
+	enc, _ := NewWriter(ioutil.Discard, WithEncoderLevel(SpeedDefault))
+	n, err := io.Copy(enc, bytes.NewBuffer(data))
+	if err != nil {
+		b.Fatal(err)
+	}
+	if n != wantSize {
+		b.Fatal(n, "!=", wantSize)
+	}
+	b.ResetTimer()
+	b.ReportAllocs()
+	b.SetBytes(wantSize)
+	for i := 0; i < b.N; i++ {
+		enc.Reset(ioutil.Discard)
+		n, err := io.Copy(enc, bytes.NewBuffer(data))
+		if err != nil {
+			b.Fatal(err)
+		}
+		if n != wantSize {
+			b.Fatal(n, "!=", wantSize)
+		}
+	}
+}